refactor: Location 클래스 사용하도록 변경
