+++ BE/iDrop/src/main/java/ifive/idrop/websocket/direction/NaverDirectionFinder.java
@@ -2,6 +2,7 @@ package ifive.idrop.websocket.direction;
 
 import ifive.idrop.websocket.direction.dto.Direction;
 import ifive.idrop.websocket.direction.dto.NaverDirectionResponse;
+import ifive.idrop.websocket.location.dto.Location;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Value;
@@ -26,8 +27,8 @@ public class NaverDirectionFinder {
 
     private final RestTemplate restTemplate = new RestTemplate();
 
-    public Direction getDirection(Double startLongitude, Double startLatitude, Double endLongitude, Double endLatitude) throws Exception {
-        String url = makeUrl(startLongitude, startLatitude, endLongitude, endLatitude);
+    public Direction getDirection(Location startLocation, Location endLocation) throws Exception {
+        String url = makeUrl(startLocation, endLocation);
         HttpHeaders headers = makeHeader();
 
         RequestEntity<Void> request = RequestEntity
@@ -39,10 +40,10 @@ public class NaverDirectionFinder {
         return new Direction(response.getPath());
     }
 
-    private String makeUrl(Double startLongitude, Double startLatitude, Double endLongitude, Double endLatitude) {
+    private String makeUrl(Location startLocation, Location endLocation) {
         StringBuilder sb = new StringBuilder(NAVER_DIRECTION_URL);
-        sb.append("?start=").append(startLongitude).append(",").append(startLatitude);
-        sb.append("&goal=").append(endLongitude).append(",").append(endLatitude);
+        sb.append("?start=").append(startLocation.getLongitude()).append(",").append(startLocation.getLatitude());
+        sb.append("&goal=").append(endLocation.getLongitude()).append(",").append(endLocation.getLatitude());
         return sb.toString();
     }
