+++ BE/iDrop/src/main/java/ifive/idrop/websocket/location/dto/Location.java
@@ -1,6 +1,5 @@
-package ifive.idrop.websocket.location;
+package ifive.idrop.websocket.location.dto;
 
-import ifive.idrop.websocket.location.dto.DriverGeoLocation;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
